requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-image>=0.18.0
matplotlib>=3.3.0
Pillow>=8.0.0
tqdm>=4.60.0
streamlit>=1.28.0
plotly>=5.15.0
